Fix sorting, remove repeated code
