+++ src/pages/patch/patchTabs/Tasks.tsx
@@ -2,9 +2,7 @@ import { useQuery } from "@apollo/client";
 import styled from "@emotion/styled";
 import Button from "@leafygreen-ui/button";
 import { Skeleton } from "antd";
-import every from "lodash.every";
-import queryString from "query-string";
-import { useParams, useLocation, useHistory } from "react-router-dom";
+import { useParams, useHistory } from "react-router-dom";
 import { usePatchAnalytics } from "analytics";
 import { PageSizeSelector } from "components/PageSizeSelector";
 import { Pagination } from "components/Pagination";
@@ -16,14 +14,11 @@ import { useToastContext } from "context/toast";
 import { PatchTasksQuery, PatchTasksQueryVariables } from "gql/generated/types";
 import { GET_PATCH_TASKS } from "gql/queries";
 import { useNetworkStatus } from "hooks";
+import { useGetTaskQueryVariables } from "hooks/useGetTaskQueryVariables";
 import { useTaskSortQueryParams } from "hooks/useTaskSortQueryParams";
 import { PatchTasksTable } from "pages/patch/patchTabs/tasks/PatchTasksTable";
 import { TaskFilters } from "pages/patch/patchTabs/tasks/TaskFilters";
-import { PatchTasksQueryParams, TaskStatus } from "types/task";
-import { url } from "utils";
-import { parseSortString } from "./util";
 
-const { getPageFromSearch, getLimitFromSearch } = url;
 interface Props {
   taskCount: number;
 }
@@ -31,14 +26,12 @@ interface Props {
 export const Tasks: React.FC<Props> = ({ taskCount }) => {
   const { id: resourceId } = useParams<{ id: string }>();
 
-  const { search } = useLocation();
   const router = useHistory();
   const patchAnalytics = usePatchAnalytics();
   const dispatchToast = useToastContext();
 
-  const queryVariables = getQueryVariables(search, resourceId);
+  const queryVariables = useGetTaskQueryVariables();
   const { limit, page } = queryVariables;
-
   const sorts = useTaskSortQueryParams();
 
   const { data, startPolling, stopPolling } = useQuery<
@@ -104,68 +97,6 @@ export const Tasks: React.FC<Props> = ({ taskCount }) => {
       )}
     </>
   );
-};
-
-const getString = (param: string | string[]): string =>
-  Array.isArray(param) ? param[0] : param;
-const getArray = (param: string | string[]): string[] =>
-  Array.isArray(param) ? param : [param];
-
-const statusesToIncludeInQuery = {
-  [TaskStatus.Dispatched]: true,
-  [TaskStatus.Failed]: true,
-  [TaskStatus.Inactive]: true,
-  [TaskStatus.SetupFailed]: true,
-  [TaskStatus.Started]: true,
-  [TaskStatus.StatusBlocked]: true,
-  [TaskStatus.Succeeded]: true,
-  [TaskStatus.SystemFailed]: true,
-  [TaskStatus.SystemTimedOut]: true,
-  [TaskStatus.SystemUnresponsive]: true,
-  [TaskStatus.TaskTimedOut]: true,
-  [TaskStatus.TestTimedOut]: true,
-  [TaskStatus.Undispatched]: true,
-  [TaskStatus.Unstarted]: true,
-  [TaskStatus.Aborted]: true,
-};
-
-const getStatuses = (rawStatuses: string[] | string): string[] => {
-  const statuses = getArray(rawStatuses).filter(
-    (status) => status in statusesToIncludeInQuery
-  );
-  if (
-    every(Object.keys(statusesToIncludeInQuery), (status) =>
-      statuses.includes(status)
-    )
-  ) {
-    // passing empty array for `All` value is also more performant for filtering on the backend as opposed to passing array of all statuses
-    return [];
-  }
-  return statuses;
-};
-
-const getQueryVariables = (
-  search: string,
-  resourceId: string
-): PatchTasksQueryVariables => {
-  const {
-    [PatchTasksQueryParams.Variant]: variant,
-    [PatchTasksQueryParams.TaskName]: taskName,
-    [PatchTasksQueryParams.Statuses]: rawStatuses,
-    [PatchTasksQueryParams.BaseStatuses]: rawBaseStatuses,
-    [PatchTasksQueryParams.Sorts]: sorts,
-  } = queryString.parse(search, { arrayFormat: "comma" });
-
-  return {
-    patchId: resourceId,
-    sorts: parseSortString(sorts),
-    variant: getString(variant),
-    taskName: getString(taskName),
-    statuses: getStatuses(rawStatuses),
-    baseStatuses: getStatuses(rawBaseStatuses),
-    page: getPageFromSearch(search),
-    limit: getLimitFromSearch(search),
-  };
 };
 
 const FlexContainer = styled.div`
